requirements.txt
streamlit
pandas
matplotlib
networkx
pyyaml
